Document label configuration for Slider
With configuring YARN node label with queues and also adding it to
slider configuration,  it is possible to deploy the slider application
components like Presto on just the 'labelled' nodes. This commit adds
this in the README. This was tested manually on a cluster. We still need
to add some integration tests for this.

SWARM-908