tests: Make path for presto-app package configurable
We need to add automation for testing Presto 0.115t version. But we do not
want to update the pom in this project to point to this version and
build against 115t(115t is not yet in maven repo anyway).

Hence we do not always want the presto-app package to be taken from
target/ build folder. So remove the hardcoded path and make it
configurable in yaml config.

SWARM-1540