tests: Fix intermittent failure in flex command test
Slider flex command intermittently adds the new 'flex'ed component on
the existing worker node. But later when it fails to deploy the new
worker on the existing worker it deploys the new one on a new slave node
as expected. Our tests should make sure that we wait till the components
are deployed on the new slave node.

SWARM-1601