Configure only the tpch connector by default
There is no guarantee that hive is pre-installed on a cluster, so just
configure tpch connector by default (this will be now in sync with
presto-admin server install behavior). Also hive connector can be cdh or
hdp, so it makes sense to not to default to any. If needed, users can
configure hive or any other connectors in their appConfig.json or via Ambari gui.

Issue #35